Game start and game end clarification
The fact that the game was ongoing and you could just keep choosing
and option to keep playing wasn't linear according to my wife, so I
added a couple of alerts at the beginning. The same goes for the game end,
there was a alert that made the last round result delayed.